requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Supabase
supabase>=2.10.0

# Auth & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-dotenv>=1.0.0

# Validation
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.2.0

# File handling
aiofiles>=24.1.0
pillow>=11.0.0

# Utilities
python-dateutil>=2.9.0
pytz>=2024.0
babel>=2.16.0
